Sync: Placement shifted to Abstracts Activity. All working now.
Sync icon put in Abstracts Activity actionbar.
