+++ Conference/res/menu/abstracts_menu.xml
@@ -1,0 +1,9 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android" >
+
+    	<item
+        android:id="@+id/updateAbs"
+        android:showAsAction="always"
+        android:icon="@drawable/ic_action_refresh"
+        android:title="@string/menu_updateAbs_text"/>
+	
+</menu>
